Move claim to assets rewards
commands become:

assets rewards claim one <… asset claim args>
assets rewards claim queue claim_wallet <walelt>
