+++ helium-wallet/src/cmd/assets/claim/queue.rs
@@ -1,10 +1,10 @@
 use crate::cmd::*;
-use helium_lib::{crons, keypair::Pubkey, token};
+use helium_lib::{keypair::Pubkey, queue, token};
 
 #[derive(Debug, Clone, clap::Args)]
 pub struct Cmd {
     #[command(subcommand)]
-    cmd: ClaimCommand,
+    cmd: Command,
 }
 
 impl Cmd {
@@ -13,12 +13,13 @@ impl Cmd {
     }
 }
 
+/// Queue claim transactions with Tuktuk
 #[derive(Debug, Clone, clap::Subcommand)]
-pub enum ClaimCommand {
+pub enum Command {
     Wallet(ClaimWalletCmd),
 }
 
-impl ClaimCommand {
+impl Command {
     pub async fn run(&self, opts: Opts) -> Result {
         match self {
             Self::Wallet(cmd) => cmd.run(opts).await,
@@ -26,7 +27,7 @@ impl ClaimCommand {
     }
 }
 
-/// Create and start a one time claim for all hotspots in a wallet using Tuktuk
+/// Create and start a one time claim for all assets in a wallet using Tuktuk
 ///
 /// The tuktuk system will fund the "claim_wallet" it uses to pay for claims
 /// with a small amount of SOL. When new hotspots are added, additional payee
@@ -59,7 +60,7 @@ impl ClaimWalletCmd {
 
         if self.info {
             let claim_wallet =
-                crons::claim_wallet::claim_wallet_key(&crons::TASK_QUEUE_ID, &wallet);
+                queue::claim_wallet::claim_wallet_key(&queue::TASK_QUEUE_ID, &wallet);
             let claim_info = json!({
                 "claim_wallet": token::balance_for_address(&client, &claim_wallet).await?,
             });
@@ -70,9 +71,9 @@ impl ClaimWalletCmd {
         let password = get_wallet_password(false)?;
         let keypair = opts.load_keypair(password.as_bytes())?;
         let transaction_opts = self.commit.transaction_opts(&client);
-        let (tx, _) = crons::claim_wallet::claim_wallet(
+        let (tx, _) = queue::claim_wallet::claim_wallet(
             &client,
-            &crons::TASK_QUEUE_ID,
+            &queue::TASK_QUEUE_ID,
             &wallet,
             &keypair,
             &transaction_opts,
