+++ helium-wallet/src/cmd/assets/rewards.rs
@@ -1,6 +1,5 @@
 use crate::cmd::*;
-use anyhow::Context;
-use helium_lib::{entity_key, reward, reward::ClaimableToken, token::TokenAmount};
+use helium_lib::{entity_key, reward, reward::ClaimableToken};
 
 #[derive(Debug, Clone, clap::Args)]
 pub struct Cmd {
@@ -16,7 +15,7 @@ impl Cmd {
 
 #[derive(Debug, Clone, clap::Subcommand)]
 pub enum RewardsCommand {
-    Claim(ClaimCmd),
+    Claim(assets::claim::Cmd),
     Pending(PendingCmd),
     Recipient(RecipientCmd),
     Lifetime(LifetimeCmd),
@@ -35,57 +34,7 @@ impl RewardsCommand {
     }
 }
 
-/// Claim rewards for a given hotspot
-#[derive(Debug, Clone, clap::Args)]
-pub struct ClaimCmd {
-    /// Token for command
-    #[clap(long, default_value_t)]
-    pub token: ClaimableToken,
-    #[clap(flatten)]
-    pub entity_key: entity_key::EncodedEntityKey,
-    /// The optional amount to claim
-    ///
-    /// If not specific the full pending amount is claimed, limited by the maximum
-    /// claim amount for the subdao
-    pub amount: Option<f64>,
-    /// Commit the claim transaction.
-    #[command(flatten)]
-    pub commit: CommitOpts,
-}
-
-impl ClaimCmd {
-    pub async fn run(&self, opts: Opts) -> Result {
-        let password = get_wallet_password(false)?;
-        let keypair = opts.load_keypair(password.as_bytes())?;
-        let client = opts.client()?;
-        let transaction_opts = self.commit.transaction_opts(&client);
-
-        let token_amount = self
-            .amount
-            .map(|amount| TokenAmount::from_f64(self.token, amount).amount);
-        let Some((tx, _)) = reward::claim(
-            &client,
-            self.token,
-            token_amount,
-            &self.entity_key,
-            &keypair,
-            &transaction_opts,
-        )
-        .await?
-        else {
-            bail!("No rewards to claim")
-        };
-
-        let claim_response = self
-            .commit
-            .maybe_commit(tx, &client)
-            .await
-            .context("while claiming rewards")?;
-        print_json(&claim_response.to_json())
-    }
-}
-
-/// Get or set the recipient for rewards
+/// Get or set the the recipient for rewards
 #[derive(Debug, Clone, clap::Args)]
 pub struct RecipientCmd {
     /// Token for command
