+++ helium-wallet/src/cmd/mod.rs
@@ -25,7 +25,6 @@ use std::{
 pub mod assets;
 pub mod balance;
 pub mod burn;
-pub mod claim;
 pub mod create;
 pub mod dc;
 pub mod export;
